requirements.txt
python-docx
# optional for PDF export
docx2pdf
